move modules to a dedicated conf module
